requirements.txt
# Add all needed packages for the application to run

pandas==2.1.1
requests==2.31.0
urllib3==2.0.5
numpy==1.26.0
mysql-connector-python==8.1.0
fastapi==0.103.2
